requirements.txt
databases
fastapi
httpx
pythonrestsdk
python-dotenv
uvicorn
